Revert deprecation of the 'list' option
The list option was accidentally deprecated, because I thought infinite
scrolling was implemented both in the Community Console and TW. However,
that was only implemented in the CC and not in TW, where the feature is
still being added by this extension.

This commit, therefore, readds the list option, which should be enabled
by default, except if the user disabled it before the option was
deprecated by accident.

Change-Id: Ibdaa109ca071105b060c63e848867381618288a1