[Dark Theme] Improve thread list backgrounds
Selected threads now have a background color which can be better
distinguished from read threads.

Also, a bug has been fixed which consisted in the fact that threads
which had been read and checked at the same time had the background for
read threads instead of the background for checked threads.

Fixes: #15

Change-Id: Ibb577553e8a77e1b131e20afd637a041f69880ba