Fix: support secondary Google accounts
When multiple Google accounts are signed in, Google adds an authuser
parameter in the URLs to indicate which account should be used.

This extension adds support for secondary Google Accounts, so the
extension works well when using it with them, by adding the authuser
parameter whenever it is needed.

Before this change, some features were broken if being used with a
secondary account.

Note that this doesn't fix the CC bug which consists in the fact that
URLs formed by the CC don't include the authuser parameter, because this
change only fixes the URLs injected by the extension, and the HTTP
requests made by the extension.

Change-Id: I0da5911f4b41057aa94b2834bd9a4313919fcbd4