Fix: profile indicator could not retrieve number of posts
A change has been made by Google to the Community Console API, and now
the ViewUser method doesn't return the user activity summary (the
number of posts/answers count aggregated by month) if it is not
explicitly set via JSONBody[1][4][20] = true (this is the
'ViewUser.ViewOptions.withUserActivity' field).

This change adds this field to the ViewUser request so the profile
indicator continues to work.

Change-Id: I132d79d3651cee26c09bc47bd5f0f95446d25419