Remove chrome_style & browser_style manifest keys
The chrome_style CSS stylesheet applied by Chrome to the options page is
now included directly in the extension's bundle, and injected into the
options page.

This change is done in preparation for Manifest Version 3, which doesn't
allow setting chrome_style to true:
https://source.chromium.org/chromium/chromium/src/+/master:extensions/common/api/extensions_manifest_types.json;l=94;drc=149f67d78acea93b3d4f28bcd8e594301e1fdb08

This change also moves the options resources to their own folder.

Change-Id: I88d74989b38053d2e137ff166ddcdc4e00489428