[Profile indicator] Fix: sometimes it doesn't get injected
In the Community Console, check if the OP's username has already been
added to the website before starting the mutationObserver.

This is because sometimes Chrome injects the profile indicator content
script after all the page has been formed, and this means in these cases
the mutationObserver doesn't detect the username being added.

Bug: #21
Change-Id: I0158fd36c590fe26d658e3ec5cfc46f45b5c5284