[Previous posts] Improve reliability in the CC
Previously, we watched via a mutationObserver the addition of the
container inside the ec-user element.

However, the username is added to the container after the container is
added to the document, so this means it is very unreliable to query the
username just after detecting the addition of the container.

This change fixes this by watching for the addition of the username span
element instead.

Bug: #21
Change-Id: I04673d16c86c0a8677a55d4416ab4900cafdf15a