Fix: ignore the initial empty ec-app
The initial CC HTML contains an empty ec-app element, and the script
sometimes tried to initialize the intersection observer with this empty
ec-app element.

Now if the ec-app element doesn't contain the |.scrollable-content| div
(which means it's empty) the intersection observer is not initialized.

This ec-app element is always replaced later on with another ec-app
element which does contain the |.scrollable-content| div.

Bug: #21
Change-Id: I2dc54d0edf4a010a3860ffa1bdb104c8303de911