Delete <if> build flags in chrome_style.css
Those build flags were in the file because chrome_styles.css is a copy
of the file taken from the Chromium codebase.

In a Chrome build, grit is responsible for including the contents inside
the <if>...</if> tags or not depending on the build flags, but in our
case we don't preprocess those tags, and that led to an incorrect CSS
file.

Change-Id: If90a8ad470e1f524296122cc27803f20c0b4c87e