Dark theme: fix right controls' dialogs position
Allowing the right controls div to have unlimited width, in large
screens the dialogs which are shown when clicking the buttons aren't
shown below the buttons but displaced to the left.

This change sets a fixed maximum width so the dialogs are shown
correctly but there is still space for all the buttons.

Change-Id: I0499662f0accc5d3c0423d5e077d69ccceb2f4e4