Move right controls fix to console.css
The fix for the right controls which was supposed to leave sufficient
space to the dark theme switch in the header was only included in
ccdarktheme.css, so it was only applied when the dark theme was on.

This change moves the code to console.css so the fix is applied also
when the dark theme is disabled but the switch is still shown.

Note that this fix is also applied now even if the dark theme option
isn't enabled.

Change-Id: Ic6f969ed0292f86a55597b6154694c984ae74011