Dark theme: adapt to the new MD icons experiment
There is a new experiment in the Community Console called
SMEI_GOOGLE_MATERIAL_ICONS, which introduces a new design for the
icons and changes the way of styling the checkboxes in the thread list
view, which breaks the dark theme.

This change fixes the dark theme styles for those checkboxes, and adapts
their style to match the new behavior. This change is
backwards-compatible, so the dark theme won't break for folks who don't
have the experiment enabled (there is a small subtle change which is
applied in both cases: the background of filled checkboxes has been made
lighter so it contrasts better with the background and looks nicer).

This experiment is already enabled for some users, and so this was
reported by one of those users (see
pekb/thread/114559215?msgid=116471244).

Bug: 1
Change-Id: I9c8757c84311946a2a69495f64963fd4fbfed5cd