threadListAvatars: dynamic unauthorized forums
This is an implementation of point 3 in the "Idea" section of the
following doc: go/eu7T9m (public link available in the linked bug).

Paraphrasing the previous document:
This change adds code to maintain a list of forums from which the
extension couldn't load any thread, which probably correspond to private
forums. When a forum is in this list, avatars are no longer retrieved
from threads in that forum until the entry expires.

Bug: 2
Change-Id: I05e7b02818181f410855948e1af6ec75fc7c792b