Thread list avatars: order avatars in the natural direction
For left-to-right users, avatars were shown from right to left (and for
RTL users they were shown LTR). In order to improve clarity, now they
will be ordered in the natural direction, which means people who have
replied first will be shown first instead of last.

Thanks Ari for the feedback which led to this change!

Change-Id: Ia634cac85234a91481f2d27902830afdb2b433df