Dark theme: improve form controls styles
Due to the redesign of the checkboxes in the Community Console made by
Google, we previously introduced new styles to the dark theme for the
checkboxes in thread lists. However, the checkbox redesign affected not
only those checkboxes, but checkboxes throughout the UI.

This change adds styles for all checkboxes following the redesign, and
removes some checkbox styles applied only to some parts of the UI
(therefore streamlining the checkbox appearance).

It also changes styles for radio buttons and adds styles for switches.

Change-Id: I15c64809f064f9db69f08beb72951db2be6920ae