Improve API error messages with known error codes
When an API call fails, sometimes a numeric code is returned indicating
a type of generic error. These errors are known, so this change returns
the message instead of the numeric code when it is known.

This is particularly useful for the "batch lock" feature, since this
changes the error message "Error code: 7" to "Error code:
PERMISSION_DENIED", which is much clearer to users.

Change-Id: I7eb98a0442f63ea5aa35886453df5ea6e524af99