Improve sort detection
Before, we relied on the startup data object to detect whether the sort
options were compatible with this feature or not, which gave a wrong
veredict if those options were changed without reloading the page.

This change detects those options much more reliably, as it now parses
the last ViewForum request body to search for those settings.

Bug: twpowertools:5, twpowertools:22
Change-Id: Ic00200f51b611197c0158f9497597af47bee9581