autoRefreshList: ignore unrelated ViewForum requests
The XHR interceptor sends events for all ViewForum requests. This
includes the requests made by the "Mark as Duplicate" dialog and the
chat feature, which are treated as if a new thread list was loaded.

This change ignores those, and also ignores requests to load more
threads in the current thread list, since the only type of request
we're interested in is the first request to load the thread list.

Bug: twpowertools:24
Change-Id: I3be18615405f90c80ce32f9a5b72b36b506ac4e9