autoRefreshList: improve filter detection & remove first call
Before, we constructed the filter via an approximate method, which
wouldn't always get the correct filter that the Community Console was
using. Now that we're intercepting the first call to ViewForum, this
change gets the filter directly from the request (and also the forumId
value, which we always set as '0' before).

Also, we used to perform a first call to get the timestamp of the first
thread being shown. Now, we intercept the ViewForum response to get this
timestamp instead of performing a separate call to the API.

Note that sometimes when the thread list is shown the CC doesn't call
the API, but insteads gets the threads from a previous call. The only
case I've detected is when a user navigates back to a thread list, and
in this case the feature uses the previously saved data from the last
interception, which works fine. Also, if a thread list is loaded and we
navigate forward to another thread list, we can't go back to the
previous thread list (there isn't a button in the UI and the browser
button goes to the page opened before the Community Console), so the
feature doesn't break this way.

However, if there are other cases in which the Community Console doesn't
call the API neither, these might break the feature. Anyways, this
change is definitely an improvement over the old method.

Fixed: twpowertools:22
Change-Id: I66ff3b78e2089dc1decd35e6a89b208446409b86