autoRefreshList: add status indicator
In order to clarify when the feature is watching for updates in the
current thread list or not depending in the current sort options, this
change adds an indicator which states this.

It also adds styles for a "muted" state of the indicator which might
indicate in the future that the user has disabled it manually.

Apart from this, this also changes the option text to improve it.

Fixed: twpowertools:5
Change-Id: I77a65fa9091d6c9e18f09a3798eeb5197ced8647