threadListAvatars: improve private thread indicator
Previously, the private thread indicator consisted in a "key" icon, and
according to one PE this led to confusion since it could be
misinterpreted as a lock icon.

This change replaces the icon by a "crossed person" icon, and adds a
tooltip to the icon which explains why that icon is shown.

Fixed: twpowertools:43
Change-Id: I2ec28114d93bf78be1b5f9999e65f916d7b2703f