Build: correct locale folders generated by Pontoon
Pontoon (the software we use to localize the extension) generates
folders with locale codes which are not the ones used by Chrome nor
Firefox. This change adds logic to the build process which changes these
locale codes to the correct ones, so browsers can recognize and use all
translations.

For now, we're transforming 'pt-rBR' to 'pt', since for now it is the
only locale with a wrong code. We use 'pt' instead of 'pt_BR' since we
want to use its strings for both Portuguese (Portugal) and Portuguese
(Brazil).

Bug: twpowertools:40
Change-Id: If9ab8fc9646aa5252d3637fd8cf0fcdba11dffee