Options: make docs links point to the appropriate version
Until now the help button pointed to the latest version of the
"features" document, which might contain information about features
which haven't been launched yet. This change adapts this link and the
link to the "OP indicator" documentation so they point to the document
corresponding to the version installed (or the latest version if the
extension has been loaded in developer mode).

Change-Id: I2a99886b67d11da00214600620bd26f28ff1bdce