Make some of the options dynamic
This change modifies the logic of several features so they aren't
enabled/disabled depending on the options state when the page is loaded
but dynamically.

So, for instance, when the thread list avatars feature is switched from
enabled to disabled, when browsing the Community Console, newly loaded
thread lists won't have the avatars, without having to reload the whole
Community Console.

This will make "kill switches" more effective, since they will be able
to take effect without having to reload the Community Console page.

The options which still haven't been made dynamic are features which add
CSS tweaks to the Community Console. For those features (like the dark
mode) a future CL will make them dynamic.

Bug: twpowertools:61
Change-Id: I72b511dd3b2622a2e9c633850e29806953e4b284