Build options page programatically
Until now we manually created the options page layout, but we want to
add custom elements before each option (for the kill switch
functionality), and this is very cumbersome to do it manually. Thus,
this CL builds the options page layout programatically to be able to
easilly inject a custom component before each option.

Change-Id: Ib110679971fa70c9933be911c4750b7fafa1d40e