Fix profile indicator when OP has special chars
When an OP has "&" in their username, this character will be transformed
into "&amp;" in the escaped username instead of remaining invariant.
This change fixes this.

A fix was released for the same bug in the "previous posts" feature (see
Github issue #17), but we didn't notice that the same bug was present in
the "profile indicator" feature.

Change-Id: I06c7a8406d7f5d59d9404dca448a61fba7840b04