Enable "threadall" by default instead of "thread"
"thread" enables infinite scroll in threads by loading batches of
messages while scrolling down, which in my opinion is a bad user
experience. On the contrary, "threadall" enables infinite scroll in
threads by loading all the messages at once when getting to the bottom,
which again IMO is much more useful.

Thus, this change makes the "threadall" a default option instead of
"thread".

Change-Id: Ieb425cc265bf90bcec4601a0b54c7d9dd6600626