Change isReleaseVersion to isProdVersion
The experiments button in the options page was only shown on released
builds, and the link to the feature doc included a tag with the version
installed also on released builds. However, this doesn't work well in
Canary builds, since they should be treated like the unreleased
versions.

This CL fixes this by changing the function name to isProdVersion and
making the Canary builds behave like unreleased builds.

Change-Id: Ied55cf92b2317602adbe3c8dfc7f815b4c88fa19