Improve message states wording
The previous message states strings were based in the codenames for each
state, since we don't have their descriptions available and we had to
guess what they meant. This resulted in very confusing strings.

This CL fixes this issue by improving them and adding a description of
what we think each state describes in the string description, so
translators know what each state means. A possible improvement in the
future would be to show a tooltip with this description when hovering
the state chip.

Bug: twpowertools:113
Change-Id: I7538cc549abac1e001d935a9c344d943da780030