Update CWS credentials
The OAuth OOB flow will be deprecated on October 3, 2022, so this CL
adds regenerated credentials which use another flow to avoid the
deprecation.

It also adds a suffix to the name of the secret to avoid name collision,
since Zuul requires credential names to be unique throughout each
tenant.

Finally, it adds the .zuul.yaml file to the list of files which trigger
the infinitegforums-test-nightly-build job.

Fixed: twpowertools:121
Change-Id: I45e5fd4d2189a11bc9ddc5eb0525526be1118ac3