Add option to choose the thread page design in the CC
This feature allows users to choose whether to show the old thread page
design or the new one.

This CL also adds an experimental feature to activate the nested replies
experiment in the Community Console (only available in the Canary
channel). It has been added as an experimental feature because the
nested replies feature doesn't work yet in the Community Console.

Change-Id: Ifc371848b55f450bedf22d8e35fb4348fffa6c63