Fix XHR interceptor in CC RCE thread page
The XHR interceptor stopped working in the redesigned Community Console
thread pages (aka RCE/interop thread page). This is due to the fact that
the Javascript loaded before the XHR interceptor was set, and because
the requests differed in the Javascript types used, and the format of
the requests (an array to encode protobuf messages is used instead of
objects in the RCE thread page).

This CL fixes these issues.

Change-Id: I591d58f5f597a71a2794ee59b2c5cb9dd88cca9f