Fix avatars feature for threads with nested replies
The avatars feature didn't work well when a thread had nested replies.
This CL adapts the feature so it takes into account nested replies. The
order of the avatars tries to take into account when a user joined the
thread, instead of showing the avatars in the order they are shown
inside a thread.

Fixed: twpowertools:142
Change-Id: I38fdaf3c8f73787d246b0631776d91f470c12579