infiniteScroll: fix incompatibility w/ message links
When opening a link to a specific message inside a thread in both TW
basic and the Community Console, the "load more" button might be visible
for an instant before/while scrolling to and focusing the specific
message being linked. This causes the "load more" button to be pressed
automatically when the infinite scroll feature is enabled.

This CL fixes this issue by adding a delay of 3.5s before enabling the
infinite scroll feature when the URL specifies a message to focus.

Fixed: twpowertools:141
Change-Id: Ie9126141ee9f763707bd61b272ec53b4e1d8079f