Add workflow menu to thread lists
This CL adds a MD3 menu button to the bulk actions bar in thread lists.
The menu will display the available workflows, but as of now it just
contains random items, and clicking the items doesn't work (this will be
implemented later).

This implementation has been done using the newly added @material/web
package, so the elix package has been removed from the project. This is
because @material/web also uses Material Design like the Community
Console.

Bug: twpowertools:74
Change-Id: Ifb3712c7afc024e21c7ff83fb77658cd8d08bc8a