Show workflows in thread list menu
This CL adds the logic for displaying the list of workflows in the menu
which is added to thread lists. Before, the list included some fake
names. It also adds a button to manage workflows.

The next step is to create the components/logic which will allow the
user to execute the workflow in the selected threads.

Bug: twpowertools:74
Change-Id: I22d0be8a101f9e167b9408bb6046299f3bd3c787