Flatten threads: remove further info about parents
This CL removes information from the ViewThread and CreateMessage
responses which indicate that the nested messages have a parent message.

This is done because although this information is not used now to render
the replies, it might be used in the future by the Community Console
code and this might break the feature.

Bug: twpowertools:156

Change-Id: Ie32ce2b112f15395c2f15d5d6b9b5cde4cc64768