Flatten threads: remove the additional info element in the edit box
The flatten threads feature needs to inject an element with additional
information about the message to the message payload, which is hidden
via an inline CSS style. However, when the Community Console renders the
edit box, this style is removed and the additional information is
visible in the reply box.

This CL fixes this by removing this element from the reply box when
found.

Bug: twpowertools:156
Change-Id: I9eb5a965926fef91374396304e12837d5ef5c798