Tempatively fix XHRProxy bug
The XHRProxy calls the optionsWatcher server as soon as it is loaded,
but the server isn't guaranteed to be loaded and thus the request can
get lost, which breaks the XHRProxy and causes all XHR requests to fail.

This CL fixes this by ensuring the optionsWatcher server is loaded
before injecting the XHRProxy to fix this issue.

Bug: twpowertools:161

Change-Id: I12ec2d018c9fa7791053807cad19a547564e2ee2