Show an extension badge when a kill switch is active
Kill switches allow the extension maintainer to remotely disable some
features in case they start working in an undiserable manner (for
instance when Google updates the Community Console and the extension
breaks the CC completely).

This change makes it clear whether a kill switch is active by adding a
badge to the extension icon.

Change-Id: I5fbbbc33aa78c613d91625ba0ea10deb2d478a83