Add XHR proxy kill switch
Since the XHRProxy loads very early to be able to catch all requests,
we cache the kill switch value in the localStorage and this cached
value is consulted early in the page load to determine whether the
XHR proxy should be put in place.

Thus, this kill switch only takes effect after a page reload.

Fixed: twpowertools:162
Change-Id: I43a163c506657d5ba9fb4ec268b2d4409d6401e3