refactor: change info logs to debug or warning
This CL converts some info-level messages to the debug level because
they're irrelevant and shouldn't be shown by default in the Javascript
console.

Also, one info-level message has been converted to a warning because it
represents a situation that couldn't be handled by the extension which
causes the feature to behave in an unexpected manner.

Change-Id: If1a153320c502bd5a2805597c8558456039a89a1