fix(flatten-threads): handle |#action=reply| correctly
The action to open the reply editor is usually handled by the Community
Console. With the flatten threads enabled, this sometimes opens a reply
editor corresponding to a nested reply which when used wouldn't post the
reply correctly.

Thus, this CL adds logic to handle the |#action=reply| ourselves to open
the reply editor adequately, as if the user clicked our own "Reply"
button. This opens the reply editor corresponding to the parent reply
(the first reply in the reply chain).

Fixed: twpowertools:180
Change-Id: I1d734cfe0e28971939d292121ad6144f9e9f7a9a