fix(workflows): prevent dialogs from overlapping sidebar
When the "fix thread actions" option is enabled, we apply some styling
which makes the workflow dialogs overlap the sidebar/header.

This CL makes the sidebar appear behind the dialog scrim by changing the
z-index of the thread actions toolbar. The header still appears appears
above the scrim, but it's a tradeoff because otherwise the toolbar would
appear above the header when scrolling down and that's worse.

Fixed: twpowertools:150
Change-Id: I3fae699dcd1c2391ba5235f781cc5af37ef5e47c