fix(xhr-proxy): forward missing responseURL property
This change fixes a current issue with the Community Console.

The Community Console code has been updated, and it now tries to access
the responseURL property of XMLHttpRequest instances. Unfortunately,
this property was not amongst the ones that our XHRProxyInstance
forwarded from the original XMLHttpRequest instances.

Due to this the Community Console received an undefined value, making it
throw an error due to this unexpected response, and making it stop
handling the response.

Bug: twpowertools:187
Fixed: twpowertools:188
Change-Id: Ice9c66c0c4a1cc5822c24103f6837511052480fd