fix(kill-switch): handle non-semver extension versions well
This change makes the extension convert non-semver versions to semver
versions so the kill switch logic can correctly compare the extension
version with the affected versions range set in kill switches.

It does so by only allowing version numbers to have 3 parts, and
stripping further sub-versioning. In the canary version, version numbers
consist of 4 parts (the latest version number followed by a dot and the
number of commits made since that release). Thus, this fixes an issue in
the kill switch logic for the canary channel.

Fixed: twpowertools:189
Change-Id: Iee62f75c38e73b39786a0fc05de6e827edd4c8a2