chore: use mv3 in the stable channel
This CL makes the extension use Manifest Version 3 in the stable channel
for Chrome.

The MV3 builds are mature enough, since the extension has been built
with MV3 in the beta and canary channels for a long time. And this
change is necessary since we're near the MV2 deadline.

Fixed: twpowertools:83
Change-Id: I76db6ba892ca8e0dfafcad5772747f1bc78b5e3b