feat: add script to output options prototype JSON
We now define the options via Typescript instead of a JSON file, but we
need the options prototype in JSON format so we can sync new features to
the kill switch server.

Thus, this CL adds a script which outputs this data, which can be used
with the syncfeatures binary. It also updates the docs on how to sync
the features.

Change-Id: I238e4521f00e4b5602712fa8d2b4360a1f7fb827