feat(cc-slowness-fix): remove unused large data from some requests
The ViewForum and ViewThread requests may contain the
User.ProfileAbuse.account_reported_abuse field in their responses which
for some users contains megabytes of data and is seemingly unused by the
Community Console.

Thus, this CL removes this data from the responses when the feature
which fixes the CC slowness and high memory usage is enabled.

Change-Id: Idef49285c29d256c5413971c80581ba4ad65cb98