fix: tentatively fix flaky tests
In CI the tests sometimes fail with the following error:
`ReferenceError: exports is not defined`.[1]

This commit changes the ts-node Jest preset to transpile code into
CommonJS instead of using ESM, since the latter is considered
experimental. It is unknown whether this will fix the issue in the CI,
but it might help.

Due to this change, this commit also migrates a Javascript test to
Typescript.

[1]: https://zuul.corp.avm99963.com/t/gerrit/build/08e3a3723c03477cbd3f7c63d0568be9

Change-Id: I31f4e0fa6ec5d28ce97fae108e51eccc412c4535