refactor(mw-cs-bridge): convert all MW/CS bridges to Typescript
This relands commit c7d6a02ab42de23079e03bb31c5f257fe2d7f180 with
additional modifications to fix the regressions introduced by it.

Mainly, in this commit we also convert all the bridge implementations to
Typescript to ease the migration.

Otherwise we had the base classes in Typescript, the bridge
implementations in Javascript, and the users of the bridges in
Typescipt. Converting everything to Typescript prevents us from
introducing more bugs.

Bug: twpowertools:230
Change-Id: I6a6a69648a22ba9d54522d2432f66a6dd44276e9