feat(bulk-move): retrieve forum configuration from startup data
This commit changes the mock data by real forum configuration, so the
modal shows the actual list of forums, etc.

The startup data contains the forum configuration for the Community
Console's UI language, but configurations for other languages are
missing. Thus, only one language is shown in the modal.

A follow-up commit will add support for other languages.

Bug: twpowertools:51
Change-Id: I6a6a6964762249684a6fa1a2fcdea35209dbfa83