feat(bulk-move): persist details selection when switching languages
When switching languages, the selections for the details which also
exist in the new language are now kept so the user doesn't have to
select them again.

After taking a quick look, it seems like all languages contain the same
details (with different translations). This is why this behavior will be
useful.

This might also enable a "hack" for users: select your native language,
then select the detail values, and then select the target language
(where you might not understand the detail values, but the selection
will be persisted so there's no need to translate them).

The test file has been changed to be able to verify this change.

Bug: twpowertools:51
Change-Id: I6a6a6964b7f76665f0985e8cd652c3eaf861fe60