feat(bulk-move): load all language configurations
When selecting a language in the bulk move modal, now a call to the
ViewForum endpoint is made to load all the language configurations.

This will enable showing the different languages in the selector along
with the corresponding category/details configuration for all the
languages where the forum is available.

Bug: twpowertools:51
Change-Id: I6a6a6964c8fa3e48bb7a1bf8d5e1b1909c529d45