fix(bulk-move): show id in the dropdown when the name is not set
In the Google Assistant forum, the category with ID "io19" doesn't have
a name set, and so the option appears blank in the modal.

This commit thus changes all the selectors in the modal (including the
category selector) to show the ID when the name is not set.

This is how the vanilla "Move question" modal works like in the
Community Console.

Bug: twpowertools:51
Change-Id: I6a6a6964ca860028cebad5493614c8d26de35dbe