fix(bulk-move): always show forum name in the display language
The forum configuration is split into a configuration for each language
(called the "language configuration"). Thus, in order to obtain a unique
representation of the forum name, we try to do it from the configuration
corresponding to the display language configured by the user.

However, the logic for obtaining this language configuration was buggy,
because we were comparing the user's display language with the main
language set in each language configuration. This meant that when the
user set a secondary language (like es-419) as the display language,
sometimes no language configuration would match, and we would use the
first one in the list as a fallback.

This CL fixes this by checking instead whether the user's display
language is included in the supported languages list of the language
configuration, which includes the full list of languages for which that
configuration applies.

Bug: twpowertools:243, twpowertools:51
Change-Id: I6a6a6964f8be48c2869e0f0a2b49019c5e67f79c